chore: remove designate API for event and merge with project
