+++ providers/index.js
@@ -1,8 +1,4 @@
 const {
-  issueCreationCallback,
-  issueCreationAuth,
-} = require("./event-github/auth.js");
-const {
   githubAuth,
   githubAuthCallback,
   githubApp,
@@ -15,6 +11,4 @@ module.exports = {
   githubApp,
   gitlabAuth,
   gitlabAuthCallback,
-  issueCreationAuth,
-  issueCreationCallback,
 };
